+++ AppInspector.Tests/RuleProcessor/SubstringWithIndexTests.cs
@@ -14,146 +14,6 @@ namespace AppInspector.Tests.RuleProcessor;
 [ExcludeFromCodeCoverage]
 public class SubstringWithIndexTests
 {
-    private const string jsonAndXmlStringRule = @"[
-        {
-            ""id"": ""SA000005"",
-            ""name"": ""Testing.Rules.JSONandXML"",
-            ""tags"": [
-                ""Testing.Rules.JSON.JSONandXML""
-            ],
-            ""severity"": ""Critical"",
-            ""description"": ""This rule finds books from the JSON or XML titled with Franklin."",
-            ""patterns"": [
-                {
-                    ""pattern"": ""Franklin"",
-                    ""type"": ""string"",
-                    ""confidence"": ""High"",
-                    ""scopes"": [
-                        ""code""
-                    ],
-                    ""jsonpaths"" : [""$.books[*].title""],
-                    ""xpaths"" : [""/bookstore/book/title""]
-                }
-            ],
-            ""_comment"": """"
-        }
-    ]";
-
-    private const string jsonStringRule = @"[
-        {
-            ""id"": ""SA000005"",
-            ""name"": ""Testing.Rules.JSON"",
-            ""tags"": [
-                ""Testing.Rules.JSON""
-            ],
-            ""severity"": ""Critical"",
-            ""description"": ""This rule finds books from the JSON titled with Franklin."",
-            ""patterns"": [
-                {
-                    ""pattern"": ""Franklin"",
-                    ""type"": ""string"",
-                    ""confidence"": ""High"",
-                    ""scopes"": [
-                        ""code""
-                    ],
-                    ""jsonpaths"" : [""$.books[*].title""]
-                }
-            ],
-            ""_comment"": """"
-        }
-    ]";
-
-    private const string xmlStringRule = @"[
-    {
-        ""id"": ""SA000005"",
-        ""name"": ""Testing.Rules.XML"",
-        ""tags"": [
-            ""Testing.Rules.XML""
-        ],
-        ""severity"": ""Critical"",
-        ""description"": ""This rule finds books from the XML titled with Franklin."",
-        ""patterns"": [
-            {
-                ""pattern"": ""Franklin"",
-                ""type"": ""string"",
-                ""confidence"": ""High"",
-                ""scopes"": [
-                    ""code""
-                ],
-                ""xpaths"" : [""/bookstore/book/title""]
-            }
-        ],
-        ""_comment"": """"
-    }
-]";
-
-    private const string jsonData =
-        @"{
-    ""books"":
-    [
-        {
-            ""category"": ""fiction"",
-            ""title"" : ""A Wild Sheep Chase"",
-            ""author"" : ""Haruki Murakami"",
-            ""price"" : 22.72
-        },
-        {
-            ""category"": ""fiction"",
-            ""title"" : ""The Night Watch"",
-            ""author"" : ""Sergei Lukyanenko"",
-            ""price"" : 23.58
-        },
-        {
-            ""category"": ""fiction"",
-            ""title"" : ""The Comedians"",
-            ""author"" : ""Graham Greene"",
-            ""price"" : 21.99
-        },
-        {
-            ""category"": ""memoir"",
-            ""title"" : ""The Night Watch"",
-            ""author"" : ""David Atlee Phillips"",
-            ""price"" : 260.90
-        },
-        {
-            ""category"": ""memoir"",
-            ""title"" : ""The Autobiography of Benjamin Franklin"",
-            ""author"" : ""Benjamin Franklin"",
-            ""price"" : 123.45
-        }
-    ]
-}
-";
-
-    private const string xmlData =
-        @"<?xml version=""1.0"" encoding=""utf-8"" ?>   
-  <bookstore>  
-      <book genre=""autobiography"" publicationdate=""1981-03-22"" ISBN=""1-861003-11-0"">  
-          <title>The Autobiography of Benjamin Franklin</title>  
-          <author>  
-              <first-name>Benjamin</first-name>  
-              <last-name>Franklin</last-name>  
-          </author>  
-          <price>8.99</price>  
-      </book>  
-      <book genre=""novel"" publicationdate=""1967-11-17"" ISBN=""0-201-63361-2"">  
-          <title>The Confidence Man</title>  
-          <author>  
-              <first-name>Herman</first-name>  
-              <last-name>Melville</last-name>  
-          </author>  
-          <price>11.99</price>  
-      </book>  
-      <book genre=""philosophy"" publicationdate=""1991-02-15"" ISBN=""1-861001-57-6"">  
-          <title>The Gorgias</title>  
-          <author>  
-              <name>Plato</name>  
-          </author>  
-          <price>9.99</price>  
-      </book>  
-  </bookstore>
-";
-
     private const string wordBoundaryDisabledCaseSensitive = @"[
     {
         ""id"": ""SA000005"",
@@ -259,46 +119,6 @@ raceCAR withmorestuff
 racecar withmorestuff";
 
     private readonly Microsoft.ApplicationInspector.RulesEngine.Languages _languages = new();
-
-    [DataRow(jsonStringRule)]
-    [DataRow(jsonAndXmlStringRule)]
-    [DataTestMethod]
-    public void JsonSubstringRule(string rule)
-    {
-        RuleSet rules = new();
-        rules.AddString(rule, "JsonTestRules");
-        Microsoft.ApplicationInspector.RulesEngine.RuleProcessor processor = new(rules,
-            new RuleProcessorOptions { AllowAllTagsInBuildFiles = true });
-        if (_languages.FromFileNameOut("test.json", out var info))
-        {
-            var matches = processor.AnalyzeFile(jsonData, new FileEntry("test.json", new MemoryStream()), info);
-            Assert.AreEqual(1, matches.Count);
-        }
-        else
-        {
-            Assert.Fail();
-        }
-    }
-
-    [DataRow(xmlStringRule)]
-    [DataRow(jsonAndXmlStringRule)]
-    [DataTestMethod]
-    public void XmlSubstringRule(string rule)
-    {
-        RuleSet rules = new();
-        rules.AddString(rule, "XmlTestRules");
-        Microsoft.ApplicationInspector.RulesEngine.RuleProcessor processor = new(rules,
-            new RuleProcessorOptions { AllowAllTagsInBuildFiles = true });
-        if (_languages.FromFileNameOut("test.xml", out var info))
-        {
-            var matches = processor.AnalyzeFile(xmlData, new FileEntry("test.xml", new MemoryStream()), info);
-            Assert.AreEqual(1, matches.Count);
-        }
-        else
-        {
-            Assert.Fail();
-        }
-    }
 
     [TestMethod]
     public void NoDictDataAllowed()
