+++ AppInspector.Tests/RuleProcessor/WithinClauseTests.cs
@@ -556,84 +556,6 @@ http://
         }
     }
 
-    [TestMethod]
-    public void TestXmlWithAndWithoutNamespace()
-    {
-        var content = @"<?xml version=""1.0"" encoding=""UTF-8""?>
-<project xmlns=""http://maven.apache.org/POM/4.0.0"" xmlns:xsi=""http://www.w3.org/2001/XMLSchema-instance"" xsi:schemaLocation=""http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>xxx</groupId>
-  <artifactId>xxx</artifactId>
-  <version>0.1.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-
-  <name>${project.groupId}:${project.artifactId}</name>
-  <description />
-
-  <properties>
-    <java.version>17</java.version>
-  </properties>
-
-</project>";
-        // The same as above but with no namespace specified
-        var noNamespaceContent = @"<?xml version=""1.0"" encoding=""UTF-8""?>
-<project>
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>xxx</groupId>
-  <artifactId>xxx</artifactId>
-  <version>0.1.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-
-  <name>${project.groupId}:${project.artifactId}</name>
-  <description />
-
-  <properties>
-    <java.version>17</java.version>
-  </properties>
-
-</project>";
-        var rule = @"[{
-    ""name"": ""Source code: Java 17"",
-    ""id"": ""CODEJAVA000000"",
-    ""description"": ""Java 17 maven configuration"",
-    ""applies_to_file_regex"": [
-      ""pom.xml""
-    ],
-    ""tags"": [
-      ""Code.Java.17""
-    ],
-    ""severity"": ""critical"",
-    ""patterns"": [
-      {
-        ""pattern"": ""17"",
-        ""xpaths"" : [""/*[local-name(.)='project']/*[local-name(.)='properties']/*[local-name(.)='java.version']""],
-        ""type"": ""regex"",
-        ""scopes"": [
-          ""code""
-        ],
-        ""modifiers"": [
-          ""i""
-        ],
-        ""confidence"": ""high""
-      }
-    ]
-  }]";
-        RuleSet rules = new();
-        var originalSource = "TestRules";
-        rules.AddString(rule, originalSource);
-        var analyzer = new Microsoft.ApplicationInspector.RulesEngine.RuleProcessor(rules,
-            new RuleProcessorOptions { Parallel = false, AllowAllTagsInBuildFiles = true });
-        if (_languages.FromFileNameOut("pom.xml", out var info))
-        {
-            var matches = analyzer.AnalyzeFile(content, new FileEntry("pom.xml", new MemoryStream()), info);
-            Assert.AreEqual(1, matches.Count);
-            matches = analyzer.AnalyzeFile(noNamespaceContent, new FileEntry("pom.xml", new MemoryStream()), info);
-            Assert.AreEqual(1, matches.Count);
-        }
-    }
-
     [DataRow(true, 1, new[] { 2 })]
     [DataRow(false, 1, new[] { 3 })]
     [DataTestMethod]
